Added TS and worked-example documentation, fixed tags
Removed Phil's question about how to treat worked-examples from documentation. If there is more than one worked-example, we will not treat them as separate steps.
Changed tip-for-success to tips-for-success to match PDF label.
Changed link-to-physics and links-in-physics to links-to-physics.
Added clarification to Teacher Edition tag; if TS content cannot be placed within the nearest element, it is to be placed *after* that element (and not before). This accounts for a case in which it absolutely cannot be placed before -- nothing can be placed before the chapter splash image.