change tag: ost-test-prep and ost-chapter-review
This is a new tag for Tutor to distinguish between questions that are test-prep and chapter-review. Previously these were just chapter-review and test-prep. Because these tags are NOT needed for PDF styling, and because Tutor needs to be able to distinguish between chapter-review and test-prep, and because we do not want to use "visual" (no prefix) tags for Tutor reasons, we are making this change to HS Physics.